refactor helper location, parse k/v pairs in querystring to attributes
