+++ src/main/java/io/orangebeard/listener/helper/OrangebeardPropertyHelper.java
@@ -1,9 +1,15 @@
-package io.orangebeard.listener.responders.run;
+package io.orangebeard.listener.helper;
 
 import io.orangebeard.client.OrangebeardProperty;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 import fitnesse.wiki.SystemVariableSource;
 
+import static fitnesse.http.Request.decodeContent;
+
 public class OrangebeardPropertyHelper {
+    private static final Pattern queryStringPattern = Pattern.compile("([^=&]*)=?([^&]*)&?");
 
     private OrangebeardPropertyHelper() {
     }
@@ -20,5 +26,18 @@ public class OrangebeardPropertyHelper {
 
     public static void setDescription(String description) {
         System.setProperty(OrangebeardProperty.DESCRIPTION.getPropertyName(), description);
+    }
+
+    public static void setAttributesFromQueryString(String queryString) {
+        Matcher match = queryStringPattern.matcher(queryString);
+        StringBuilder attrs = new StringBuilder();
+        while (match.find()) {
+            String key = decodeContent(match.group(1));
+            String value = decodeContent(match.group(2));
+            if (value != null && !value.isEmpty()) {
+                attrs.append(key).append(":").append(value).append(";");
+            }
+        }
+        System.setProperty(OrangebeardProperty.ATTRIBUTES.getPropertyName(), attrs.toString());
     }
 }
