+++ src/main/java/io/orangebeard/listener/responders/run/OrangebeardEnabledSuiteResponder.java
@@ -1,6 +1,7 @@
 package io.orangebeard.listener.responders.run;
 
 import io.orangebeard.listener.OrangebeardTestSystemListener;
+import io.orangebeard.listener.helper.OrangebeardPropertyHelper;
 
 import fitnesse.FitNesseContext;
 import fitnesse.http.Request;
@@ -14,6 +15,8 @@ public class OrangebeardEnabledSuiteResponder extends fitnesse.responders.run.Su
     public Response makeResponse(FitNesseContext context, Request request) throws Exception {
         OrangebeardPropertyHelper.setOrangebeardSystemProperties(context.variableSource);
         OrangebeardPropertyHelper.setTestSetName(request.getResource());
+        OrangebeardPropertyHelper.setAttributesFromQueryString(request.getQueryString());
+
         orangebeardListener = new OrangebeardTestSystemListener("orangebeard.properties", true);
 
         return super.makeResponse(context, request);
